created FactoryUtils, to share useful generator code between the different apps within Blockly Factory
