+++ demos/blocklyfactory/block_library_controller.js
@@ -60,7 +60,7 @@ BlockLibraryController = function(blockLibraryName, opt_blockLibraryStorage) {
  * @return {string} The current block's type.
  */
 BlockLibraryController.prototype.getCurrentBlockType_ = function() {
-  var rootBlock = BlockFactory.getRootBlock(BlockFactory.mainWorkspace);
+  var rootBlock = FactoryUtils.getRootBlock(BlockFactory.mainWorkspace);
   var blockType = rootBlock.getFieldValue('NAME').trim().toLowerCase();
   // Replace white space with underscores
   return blockType.replace(/\W/g, '_').replace(/^(\d)/, '_\\1');
