+++ demos/blocklyfactory/index.html
@@ -10,6 +10,7 @@
   <script src="../../blockly_uncompressed.js"></script>
   <script src="../../generators/javascript.js"></script>
   <script src="../../../closure-library/closure/goog/base.js"></script>
+  <script src="factory_utils.js"></script>
   <script src="factory.js"></script>
   <script src="block_library_view.js"></script>
   <script src="block_library_storage.js"></script>
